+++ src/frequenz/client/common/__init__.py
@@ -2,3 +2,10 @@
 # Copyright © 2023 Frequenz Energy-as-a-Service GmbH
 
 """Common code and utilities for Frequenz API clients."""
+
+from ._exception import ClientCommonError, UnspecifiedValueError
+
+__all__ = [
+    "ClientCommonError",
+    "UnspecifiedValueError",
+]