+++ src/frequenz/client/common/microgrid/proto/v1alpha8/__init__.py
@@ -3,14 +3,8 @@
 
 """Conversion of microgrid objects from/to protobuf v1alpha8."""
 
-from ._microgrid import (
-    microgrid_from_proto,
-    microgrid_status_from_proto,
-    microgrid_status_to_proto,
-)
+from ._microgrid import microgrid_from_proto
 
 __all__ = [
     "microgrid_from_proto",
-    "microgrid_status_from_proto",
-    "microgrid_status_to_proto",
 ]